Add spa rename with default
