+++ blocks/spa-output-block/output-types/spa.jsx
